requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.2.0
colorama==0.4.6
cytoolz==0.12.2
discordwebhook==1.0.3
distlib==0.3.7
emoji==2.8.0
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.3
filelock==3.12.4
frozenlist==1.4.0
hexbytes==0.3.1
identify==2.5.29
idna==3.4
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
mock==5.1.0
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
parsimonious==0.9.0
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.4.0
protobuf==4.24.3
pycryptodome==3.18.0
pyproject-api==1.6.1
pytest==7.4.2
python-dotenv==1.0.0
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.2
tomli==2.0.1
toolz==0.12.0
tox==4.11.3
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.5
web3==6.9.0
websockets==11.0.3
yarl==1.9.2